requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
bleach==2.1.1
bootstrap-admin==0.3.7.1
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.12
django-admin-bootstrapped==2.5.7
django-allauth==0.33.0
django-appconf==1.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-haystack==2.7.0
django-imagekit==4.0.1
django-import-export==1.0.0
django-redis==4.9.0
django-rest-swagger==2.2.0
django-reversion==2.0.13
django-tinymce==2.7.0
django-widgets-improved==1.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dukpy==0.2.0
et-xmlfile==1.0.1
future==0.15.2
gunicorn==19.7.1
html5lib==1.0b10
httpie==0.9.9
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
javascripthon==0.10
jdcal==1.4
jieba==0.39
Jinja2==2.10
jupyter-echarts-pypkg==0.1.1
kiwisolver==1.0.1
lml==0.0.2
lxml==4.2.3
macropy3==1.1.0b2
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.2.2
mysqlclient==1.3.12
numpy==1.14.5
oauthlib==2.0.3
odfpy==1.3.6
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.5.3
parsel==1.4.0
pilkit==2.0
Pillow==4.2.1
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.1
PyMySQL==0.7.11
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.1.0
Scrapy==1.5.0
scrapy-redis==0.6.8
service-identity==17.0.0
simplejson==3.14.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
Twisted==18.4.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
user-agent==0.1.9
w3lib==1.19.0
webencodings==0.5.1
Whoosh==2.7.4
wordcloud==1.4.1
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zope.interface==4.5.0
